+++ terraform/codebase-pipelines/codepipeline.tf
@@ -141,6 +141,30 @@ resource "aws_codepipeline" "codebase_pipeline" {
           }
         }
       }
+
+      dynamic "action" {
+        for_each = coalesce(stage.value.requires_cache_invalidation, false) ? [1] : []
+        content {
+          name             = "InvalidateCache-${stage.value.name}"
+          category         = "Build"
+          owner            = "AWS"
+          provider         = "CodeBuild"
+          input_artifacts  = ["deploy_source"]
+          output_artifacts = []
+          version          = "1"
+          run_order        = length(local.service_order_list) + 2
+
+          configuration = {
+            ProjectName = aws_codebuild_project.invalidate_cache[""].name
+            EnvironmentVariables : jsonencode([
+              { name : "CACHE_INVALIDATION_CONFIG", value : jsonencode(local.cache_invalidation_map) },
+              { name : "APPLICATION", value : var.application },
+              { name : "ENVIRONMENT", value : stage.value.name },
+              { name : "ENV_CONFIG", value : jsonencode(local.base_env_config) },
+            ])
+          }
+        }
+      }
     }
   }
 
@@ -256,6 +280,30 @@ resource "aws_codepipeline" "manual_release_pipeline" {
             { name : "REPOSITORY_URL", value : local.repository_url },
             { name : "SERVICE", value : action.value.name },
             { name : "SLACK_CHANNEL_ID", value : var.slack_channel, type : "PARAMETER_STORE" },
+          ])
+        }
+      }
+    }
+
+    dynamic "action" {
+      for_each = toset(local.cache_invalidation_enabled ? [""] : [])
+      content {
+        name             = "InvalidateCache"
+        category         = "Build"
+        owner            = "AWS"
+        provider         = "CodeBuild"
+        input_artifacts  = ["deploy_source"]
+        output_artifacts = []
+        version          = "1"
+        run_order        = length(local.service_order_list) + 2
+
+        configuration = {
+          ProjectName = aws_codebuild_project.invalidate_cache[""].name
+          EnvironmentVariables : jsonencode([
+            { name : "CACHE_INVALIDATION_CONFIG", value : jsonencode(local.cache_invalidation_map) },
+            { name : "APPLICATION", value : var.application },
+            { name : "ENVIRONMENT", value : "#{variables.ENVIRONMENT}" },
+            { name : "ENV_CONFIG", value : jsonencode(local.base_env_config) },
           ])
         }
       }