+++ terraform/codebase-pipelines/variables.tf
@@ -29,6 +29,16 @@ variable "additional_ecr_repository" {
   default = null
 }
 
+variable "cache_invalidation" {
+  type = object({
+    domains = map(object({
+      paths       = list(string)
+      environment = string
+    }))
+  })
+}
+
+
 variable "pipelines" {
   type = list(object(
     {